+++ lib/input/amqp_0_9.go
@@ -43,6 +43,9 @@ Messages consumed by this input can be processed in parallel, meaning a single
 instance of this input can utilise any number of threads within a
 ` + "`pipeline`" + ` section of a config.
 
+Use the ` + "`batching`" + ` fields to configure an optional
+[batching policy](../batching.md#batch-policy).
+
 It's possible for this input type to declare the target queue by setting
 ` + "`queue_declare.enabled` to `true`" + `, if the queue already exists then
 the declaration passively verifies that they match the target fields.